[level 1] Title: 같은 숫자는 싫어, Time: 27.63 ms, Memory: 116 MB -BaekjoonHub

+++ 프로그래머스/lv1/12906. 같은 숫자는 싫어/같은 숫자는 싫어.java
@@ -11,7 +11,7 @@ public class Solution {
         }
         int[] answer = new int[list.size()];
         for(int i = 0 ; i<list.size();i++){
-            answer[i]=list.get(i).intValue();
+            answer[i]=list.get(i);
         }
         return answer;
     }
